fixes: - new assets in svg added - bg patterns fixed - refactor mobile navigation component
