+++ src/assets/img.js
@@ -1,14 +1,18 @@
-export const img=[
-    {
-        id: 1, imgUrl: "/assets/meetups_foto.png"
-    },
-    {
-        id: 2, imgUrl: "/assets/meetups_foto.png"
-    },
-    {
-        id: 3, imgUrl: "/assets/meetups_foto.png"
-    },
-    {
-        id: 4, imgUrl: "/assets/meetups_foto.png"
-    }
-] 
+export const img = [
+  {
+    id: 1,
+    imgUrl: "/assets/meetups_foto.png",
+  },
+  {
+    id: 2,
+    imgUrl: "/assets/meetups_foto.png",
+  },
+  {
+    id: 3,
+    imgUrl: "/assets/meetups_foto.png",
+  },
+  {
+    id: 4,
+    imgUrl: "/assets/meetups_foto.png",
+  },
+];
