+++ src/components/Nav.astro
@@ -1,5 +1,5 @@
 ---
-import MobileNav from "./MobileNav";
+import MobileNav from "./mobile-nav";
 ---
 
 <nav
